--- project/family_tree_app/apps.py
@@ -1,12 +1,0 @@
-from django.apps import AppConfig
-
-class FamilyTreeConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'project'
-
-
-from django.contrib import admin
-# from .models import Member
-
-# admin.site.register(Member)
-
