--- project/family_tree_project/wsgi.py
@@ -1,16 +1,0 @@
-"""
-WSGI config for family_tree_project project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.2/howto/deployment/wsgi/
-"""
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'family_tree_project.settings')
-
-application = get_wsgi_application()
