+++ source/main/views.py
@@ -7,3 +7,5 @@ class IndexPageView(TemplateView):
 
 class ChangeLanguageView(TemplateView):
     template_name = 'main/change_language.html'
+
+
